✏️ Fix : 구독 관련 userId 자료형 변경

+++ business/src/main/java/com/mzbr/business/member/controller/MemberController.java
@@ -80,7 +80,7 @@ public class MemberController {
 	}
 
 	@PostMapping("/subscribe/{userId}")
-	public ResponseEntity<Void> subscribe(@PathVariable int userId, @AuthenticationPrincipal UserDetails userDetails) {
+	public ResponseEntity<Void> subscribe(@PathVariable long userId, @AuthenticationPrincipal UserDetails userDetails) {
 
 		memberService.subscribe(MemberSubscribeDto.of(userId, Integer.parseInt(userDetails.getUsername())));
 
@@ -88,7 +88,7 @@ public class MemberController {
 	}
 
 	@GetMapping("/subscriber/{userId}")
-	public ResponseEntity<MemberSubscribeListDto.Response> getSubscribeList(@PathVariable int userId) {
+	public ResponseEntity<MemberSubscribeListDto.Response> getSubscribeList(@PathVariable long userId) {
 
 		List<MemberDto> subscribeList = memberService.getSubscribeList(userId);
